filter_test.go
package bloom

import (
	"github.com/stretchr/testify/assert"
	"testing"
)

func BenchmarkFilter_Add(b *testing.B) {
	filter := NewFilter512()
	numEntries := 1000
	var nCollisions, firstCollision []int

	// benchmark loop
	for n := 0; n < b.N; n++ {
		nCollisions, firstCollision = benchmarkAdd(&filter, n, numEntries)
	}

	b.Logf("Filtersize: %d bits, hashes: %d\n", len(filter.fingerprint)*8, len(filter.seeds))
	b.Logf("Support: %d", b.N)
	b.Logf("First collision - ")
	statistics(firstCollision, b)
	b.Logf("Average collisions for %d entries - ", numEntries)
	statistics(nCollisions, b)
}

func TestFilter_Add(t *testing.T) {
	filter := Filter{
		fingerprint: make([]byte, 32),
		seeds:       []uint32{0, 1, 2},
	}
	filter = NewFilter512()
	d1 := generateData()
	d2 := generateData()
	assert.True(t, filter.Add(d1), "failed to add first data point")
	assert.True(t, filter.Add(d2), "failed to add second data point")
	assert.False(t, filter.Add(d2), "adding data point for the second time should fail")
}

func TestFilter_Check(t *testing.T) {
	filter := Filter{
		fingerprint: make([]byte, 32),
		seeds:       []uint32{0, 1, 2},
	}
	filter = NewFilter512()
	d1 := generateData()
	d2 := generateData()

	if !filter.Add(d1) {
		t.Errorf("Failed to add data to an empty filter")
	}

	assert.True(t, filter.Check(d1), "failed to identify only datapoint in filter")
	assert.False(t, filter.Check(d2), "identified datapoint not in filter")
}

func TestFilter_hashValues(t *testing.T) {
	// TODO
}

func TestNewFilter512(t *testing.T) {
	flt := NewFilter512()
	assert.Equal(t, 512, len(flt.fingerprint), "incorrect fingerprint length")
	assert.Equal(t, 5, len(flt.seeds), "incorrect number of hash seeds")
	assert.Equal(t, [32]byte{}, flt.checksum, "initial checksum not zero")
}
